Added SpanButton to reduce code duplication
